move copy methods to attachment utils
